backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases[sqlite]
pydantic
passlib[bcrypt]
python-jose[cryptography]
